+++ src/components/BodyTaskProgress/__tests__/index.rendering.test.tsx
